Renamed a package.
