+++ app/src/main/java/com/codefororlando/streettrees/di/TreeProviderModule.java
@@ -1,4 +1,4 @@
-package com.codefororlando.streettrees.component;
+package com.codefororlando.streettrees.di;
 
 import com.codefororlando.streettrees.api.providers.TreeProvider;
